@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class AlertPage extends StatefulWidget {
-  const AlertPage({Key key}) : super(key: key);
+  const AlertPage({Key? key}) : super(key: key);
 
   @override
   _AlertPageState createState() => _AlertPageState();
@@ -10,8 +10,6 @@ class AlertPage extends StatefulWidget {
 class _AlertPageState extends State<AlertPage> {
   @override
   Widget build(BuildContext context) {
-    return Container(
-      
-    );
+    return Container();
   }
 }
